# ap.services.user/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "APJS User Service",
    "version": "v1"
  },
  "paths": {
    "/addUser": {
      "post": {
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "user",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "firstName": {
                  "type": "string",
                  "required": true
                },
                "middleName": {
                  "type": "string",
                  "required": false
                },
                "lastName": {
                  "type": "string",
                  "required": true
                },
                "userName": {
                  "type": "string",
                  "required": true
                },
                "password": {
                  "type": "string",
                  "required": true
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "200 response"
          },
          "400": {
            "description": "400 response"
          }
        }
      }
    },
    "/getUser": {
      "get": {
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "200 response"
          },
          "400": {
            "description": "400 response"
          },
          "404": {
            "description": "200 response"
          }
        }
      }
    },
    "/getUsers": {
      "get": {
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "200 response"
          },
          "404": {
            "description": "404 response"
          }
        }
      }
    },
    "/login": {
      "get": {
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "userName",
            "in": "body",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "body",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "200 response"
          },
          "400": {
            "description": "400 response"
          },
          "404": {
            "description": "200 response"
          }
        }
      }
    },
    "/changePassword": {
      "post": {
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "id",
            "required": true,
            "schema": {
              "type": "number"
            }
          },
          {
            "in": "body",
            "name": "oldPassword",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "body",
            "name": "newPassword",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "200 response"
          },
          "400": {
            "description": "400 response"
          },
          "500": {
            "description": "500 response"
          }
        }
      }
    },
    "/changeUserState": {
      "post": {
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "id",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "200 response"
          },
          "400": {
            "description": "400 response"
          },
          "500": {
            "description": "500 response"
          }
        }
      }
    },
    "/updateUser": {
      "post": {
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "user",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "required": true
                },
                "firstName": {
                  "type": "string",
                  "required": true
                },
                "middleName": {
                  "type": "string",
                  "required": false
                },
                "lastName": {
                  "type": "string",
                  "required": true
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "200 response"
          },
          "400": {
            "description": "400 response"
          }
        }
      }
    }
  },
  "consumes": ["application/json"]
}